fix: use expected path when generating a domain
The `--directory` parameter is used to determine the path for the domain library, and was introduced in #35. This commit corrects its usage so that the `--directory` appears before the domain's name in the path instead of appearing after.

For example, this command:

ng g @angular-architects/ddd:domain user --directory=management

Will now generate the path `libs/management/user/domain` instead of `libs/user/management/domain`.

Fixes #79
Fixes #12